+++ configs/ts.js
@@ -9,20 +9,6 @@ export const tsConfig = {
     '@typescript-eslint': tsPlugin,
   },
   rules: {
-    ...[
-      'default-param-last',
-      'no-implied-eval',
-      'no-unused-vars',
-      'prefer-promise-reject-errors',
-    ].reduce(
-      (prevRules, ruleName) => ({
-        ...prevRules,
-        [ruleName]: 'off',
-        [`@typescript-eslint/${ruleName}`]: jsConfig.rules[ruleName],
-      }),
-      {}
-    ),
-
     /**
      * @see https://typescript-eslint.io/rules/adjacent-overload-signatures
      */
@@ -44,6 +30,16 @@ export const tsConfig = {
     '@typescript-eslint/ban-ts-comment': ['error'],
 
     /**
+     * @see https://typescript-eslint.io/rules/ban-tslint-comment
+     */
+    '@typescript-eslint/ban-tslint-comment': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/class-literal-property-style
+     */
+    '@typescript-eslint/class-literal-property-style': ['error'],
+
+    /**
      * @see https://typescript-eslint.io/rules/consistent-generic-constructors
      */
     '@typescript-eslint/consistent-generic-constructors': ['error'],
@@ -52,6 +48,11 @@ export const tsConfig = {
      * @see https://typescript-eslint.io/rules/consistent-indexed-object-style
      */
     '@typescript-eslint/consistent-indexed-object-style': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/consistent-type-assertions
+     */
+    '@typescript-eslint/consistent-type-assertions': ['error'],
 
     /**
      * @see https://typescript-eslint.io/rules/consistent-type-definitions
@@ -73,9 +74,27 @@ export const tsConfig = {
     }],
 
     /**
+     * @see https://typescript-eslint.io/rules/default-param-last
+     */
+    'default-param-last': ['off'],
+    '@typescript-eslint/default-param-last': jsConfig.rules['default-param-last'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/dot-notation
+     */
+    'dot-notation': ['off'],
+    '@typescript-eslint/dot-notation': jsConfig.rules['dot-notation'],
+
+    /**
      * @see https://typescript-eslint.io/rules/method-signature-style
      */
     '@typescript-eslint/method-signature-style': ['error', 'property'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/no-array-constructor
+     */
+    'no-array-constructor': ['off'],
+    '@typescript-eslint/no-array-constructor': jsConfig.rules['no-array-constructor'],
 
     /**
      * @see https://typescript-eslint.io/rules/no-array-delete
@@ -86,6 +105,16 @@ export const tsConfig = {
      * @see https://typescript-eslint.io/rules/no-base-to-string
      */
     '@typescript-eslint/no-base-to-string': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/no-confusing-non-null-assertion
+     */
+    '@typescript-eslint/no-confusing-non-null-assertion': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/no-confusing-void-expression
+     */
+    '@typescript-eslint/no-confusing-void-expression': ['error'],
 
     /**
      * @see https://typescript-eslint.io/rules/no-deprecated
@@ -108,14 +137,35 @@ export const tsConfig = {
     '@typescript-eslint/no-duplicate-type-constituents': ['error'],
 
     /**
+     * @see https://typescript-eslint.io/rules/no-dynamic-delete
+     */
+    '@typescript-eslint/no-dynamic-delete': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/no-empty-function
+     */
+    'no-empty-function': ['off'],
+    '@typescript-eslint/no-empty-function': jsConfig.rules['no-empty-function'],
+
+    /**
      * @see https://typescript-eslint.io/rules/no-empty-object-type
      */
     '@typescript-eslint/no-empty-object-type': ['error'],
 
     /**
+     * @see https://typescript-eslint.io/rules/no-explicit-any
+     */
+    '@typescript-eslint/no-explicit-any': ['off'],
+
+    /**
      * @see https://typescript-eslint.io/rules/no-extra-non-null-assertion
      */
     '@typescript-eslint/no-extra-non-null-assertion': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/no-extraneous-class
+     */
+    '@typescript-eslint/no-extraneous-class': ['error'],
 
     /**
      * @see https://typescript-eslint.io/rules/no-floating-promises
@@ -128,6 +178,27 @@ export const tsConfig = {
     '@typescript-eslint/no-for-in-array': ['error'],
 
     /**
+     * @see https://typescript-eslint.io/rules/no-implied-eval
+     */
+    'no-implied-eval': ['off'],
+    '@typescript-eslint/no-implied-eval': jsConfig.rules['no-implied-eval'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/no-inferrable-types
+     */
+    '@typescript-eslint/no-inferrable-types': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/no-invalid-void-type
+     */
+    '@typescript-eslint/no-invalid-void-type': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/no-meaningless-void-operator
+     */
+    '@typescript-eslint/no-meaningless-void-operator': ['error'],
+
+    /**
      * @see https://typescript-eslint.io/rules/no-misused-new
      */
     '@typescript-eslint/no-misused-new': ['error'],
@@ -136,6 +207,11 @@ export const tsConfig = {
      * @see https://typescript-eslint.io/rules/no-misused-promises
      */
     '@typescript-eslint/no-misused-promises': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/no-mixed-enums
+     */
+    '@typescript-eslint/no-mixed-enums': ['error'],
 
     /**
      * @see https://typescript-eslint.io/rules/no-namespace
@@ -160,6 +236,7 @@ export const tsConfig = {
     /**
      * @see https://typescript-eslint.io/rules/no-redeclare
      */
+    'no-redeclare': ['off'],
     '@typescript-eslint/no-redeclare': ['off'],
 
     /**
@@ -203,6 +280,11 @@ export const tsConfig = {
     '@typescript-eslint/no-unnecessary-template-expression': ['error'],
 
     /**
+     * @see https://typescript-eslint.io/rules/no-unnecessary-type-arguments
+     */
+    '@typescript-eslint/no-unnecessary-type-arguments': ['error'],
+
+    /**
      * @see https://typescript-eslint.io/rules/no-unnecessary-type-assertion
      */
     '@typescript-eslint/no-unnecessary-type-assertion': ['error'],
@@ -211,6 +293,11 @@ export const tsConfig = {
      * @see https://typescript-eslint.io/rules/no-unnecessary-type-constraint
      */
     '@typescript-eslint/no-unnecessary-type-constraint': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/no-unnecessary-type-parameters
+     */
+    '@typescript-eslint/no-unnecessary-type-parameters': ['error'],
 
     /**
      * @see https://typescript-eslint.io/rules/no-unsafe-argument
@@ -258,6 +345,24 @@ export const tsConfig = {
     '@typescript-eslint/no-unsafe-unary-minus': ['error'],
 
     /**
+     * @see https://typescript-eslint.io/rules/no-unused-expressions
+     */
+    'no-unused-expressions': ['off'],
+    '@typescript-eslint/no-unused-expressions': jsConfig.rules['no-unused-expressions'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/no-unused-vars
+     */
+    'no-unused-vars': ['off'],
+    '@typescript-eslint/no-unused-vars': jsConfig.rules['no-unused-vars'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/no-useless-constructor
+     */
+    'no-useless-constructor': ['off'],
+    '@typescript-eslint/no-useless-constructor': jsConfig.rules['no-useless-constructor'],
+
+    /**
      * @see https://typescript-eslint.io/rules/no-wrapper-object-types
      */
     '@typescript-eslint/no-wrapper-object-types': ['error'],
@@ -266,6 +371,12 @@ export const tsConfig = {
      * @see https://typescript-eslint.io/rules/non-nullable-type-assertion-style
      */
     '@typescript-eslint/non-nullable-type-assertion-style': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/only-throw-error
+     */
+    'no-throw-literal': ['off'],
+    '@typescript-eslint/only-throw-error': jsConfig.rules['no-throw-literal'],
 
     /**
      * @see https://typescript-eslint.io/rules/prefer-as-const
@@ -293,14 +404,56 @@ export const tsConfig = {
     '@typescript-eslint/prefer-includes': ['error'],
 
     /**
+     * @see https://typescript-eslint.io/rules/prefer-literal-enum-member
+     */
+    '@typescript-eslint/prefer-literal-enum-member': ['error'],
+
+    /**
      * @see https://typescript-eslint.io/rules/prefer-namespace-keyword
      */
     '@typescript-eslint/prefer-namespace-keyword': ['error'],
 
     /**
+     * @see https://typescript-eslint.io/rules/prefer-nullish-coalescing
+     */
+    '@typescript-eslint/prefer-nullish-coalescing': ['error'],
+
+    /**
      * @see https://typescript-eslint.io/rules/prefer-optional-chain
      */
     '@typescript-eslint/prefer-optional-chain': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/prefer-promise-reject-errors
+     */
+    'prefer-promise-reject-errors': ['off'],
+    '@typescript-eslint/prefer-promise-reject-errors': jsConfig.rules['prefer-promise-reject-errors'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/prefer-reduce-type-parameter
+     */
+    '@typescript-eslint/prefer-reduce-type-parameter': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/prefer-regexp-exec
+     */
+    '@typescript-eslint/prefer-regexp-exec': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/prefer-return-this-type
+     */
+    '@typescript-eslint/prefer-return-this-type': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/prefer-string-starts-ends-with
+     */
+    '@typescript-eslint/prefer-string-starts-ends-with': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/require-await
+     */
+    'require-await': ['off'],
+    '@typescript-eslint/require-await': jsConfig.rules['require-await'],
 
     /**
      * @see https://typescript-eslint.io/rules/restrict-plus-operands
@@ -313,6 +466,11 @@ export const tsConfig = {
     '@typescript-eslint/restrict-template-expressions': ['error'],
 
     /**
+     * @see https://typescript-eslint.io/rules/return-await
+     */
+    '@typescript-eslint/return-await': ['error'],
+
+    /**
      * @see https://typescript-eslint.io/rules/triple-slash-reference
      */
     '@typescript-eslint/triple-slash-reference': ['error'],
@@ -321,5 +479,15 @@ export const tsConfig = {
      * @see https://typescript-eslint.io/rules/unbound-method
      */
     '@typescript-eslint/unbound-method': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/unified-signatures
+     */
+    '@typescript-eslint/unified-signatures': ['error'],
+
+    /**
+     * @see https://typescript-eslint.io/rules/use-unknown-in-catch-callback-variable
+     */
+    '@typescript-eslint/use-unknown-in-catch-callback-variable': ['error'],
   },
 };